gate: Limit tenant with GBS (RunCost) and BytesOut

+++ src/apis/apps/common.go
@@ -21,7 +21,7 @@ type FunctionLimits struct {
 	Rate		uint	`json:"rate,omitempty",bson:"rate,omitempty"`
 	Burst		uint	`json:"burst,omitempty",bson:"burst,omitempty"`
 	MaxInProj	uint	`json:"maxinproj,omitempty",bson:"maxinproj,omitempty"`
-	GBS		uint64	`json:"gbs,omitempty",bson:"gbs,omitempty"`
+	GBS		float64	`json:"gbs,omitempty",bson:"gbs,omitempty"`
 	BytesOut	uint64	`json:"bytesout,omitempty",bson:"bytesout,omitempty"`
 }
